Refactor implementations
