+++ kzg_prover/serializer/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod serialize;
 pub mod type_aliases;
+pub mod implementations;
